+++ src/kit/ui/UiTheme.hpp
@@ -4,8 +4,11 @@
 
 #include <cstdio>
 #include <vita2d.h>
-#define PRIMARY_BACKGROUND_COLOR RGBA(225, 226, 225)
-#define SECONDARY_BACKGROUND_COLOR RGBA(245, 245, 246)
+
+typedef enum TypeTheme {
+    THEME_PRIMARY,
+    THEME_SECONDARY
+} TypeTheme;
 
 typedef struct{
     const char *normal;