+++ src/kit/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.5.3
+* Add TextFields class (filledDraw only)
+* Improve theme (in all ui components) to match with Material Design
+* Add textfields sample
+
 ## Version 0.5.2
 * Add Checkbox class
 * Improve Theme class